pyproject.toml
[project]
name = "friday"
version = "0.1.0"
description = "Terminal LLM assistant skeleton"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.12",
  "faststream>=0.5.0",
  "httpx>=0.27",
  "jsonschema>=4.22",
  "mcp>=1.25.0",
  "mss>=9.0",
  "python-dateutil>=2.9",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "textual>=0.58",
  "textual-dev>=1.8.0",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "taskipy>=1.12",
  "ruff>=0.6",
  "ty>=0.0",
]
voice = [
  "faster-whisper>=1.0",
  "numpy>=1.26",
  "pyttsx3>=2.90",
  "sounddevice>=0.4",
  "webrtcvad>=2.0.10",
]

[project.scripts]
friday = "friday.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "C4", "RUF"]
fixable = ["ALL"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.taskipy.tasks]
# Running
run = "uv run friday --help"

# Quality
format = "uv run ruff format ."
lint = "uv run ruff check ."
lint_fix = "uv run ruff check . --fix"
typecheck = "uv run ty check src tests"
test = "uv run pytest"

# All-in-one
check = "task format && task lint && task typecheck && task test"
